Extract WindowsSensors common base (JNA/FFM dedup)
The JNA and FFM Windows Sensors implementations duplicated the entire OHM -> LHM
-> WMI fallback orchestration, the reflection-based Libre Hardware Monitor
queries (which are backend-agnostic), and the WMI/OHM result processing (~248
duplicated lines per Sonar).

Extract a WindowsSensors base in oshi-common holding all of that. The subclasses
are reduced to four native hooks: queryOhmCpuSensor (OHM via the backend WMI
handler) and queryWmiTemperature/queryWmiFanSpeed/queryWmiVoltage (the WMI
driver queries returning the raw WmiResult, which the base processes). The OHM
CPU-identifier selection is shared via a selectOhmCpuIdentifier helper.

The subclasses drop from ~367/318 lines to ~75.
